Refactor: change status code validation
